+++ src/libvgmstream.vcxproj.filters
@@ -1729,6 +1729,9 @@
     <ClCompile Include="meta\mp4.c">
       <Filter>meta\Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="meta\mp4_faac.c">
+      <Filter>meta\Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="meta\msb_msh.c">
       <Filter>meta\Source Files</Filter>
     </ClCompile>
